requirements.txt
Django==4.1.1
djangorestframework==3.14.0
django-filter==22.1
